Moved ascension number parsing code within retry block
If some error output is being detected as the ascension number, it causes the
live_plot file to be backed up. This is not desirable, and the cause of error
is not known yet.

A possibility might be that the Flash cookie might not have been updated in time
to be ready for reading.